backprod/requirements.txt
Django==4.2.7
psycopg2-binary==2.9.9
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.0
django-cors-headers==4.3.1
Pillow==10.1.0
openpyxl==3.1.2
python-dateutil==2.8.2


#para instalar escribir en el terminal "pip install -r requirements.txt"